[refactor] Overloaded operators for AnalogExpr and Atom types inherit from AnalogExpr. Moved atom to expr to avoid circular import.
